requirements.txt
beautifulsoup4==4.9.0
bs4==0.0.1
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
cycler==0.10.0
Flask==1.1.2
Flask-WTF==0.14.3
idna==2.9
itsdangerous==1.1.0
jieba==0.42.1
Jinja2==2.11.2
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.2.1
numpy==1.18.2
Pillow==7.1.1
pyparsing==2.4.7
python-dateutil==2.8.1
requests==2.23.0
six==1.14.0
soupsieve==2.0
urllib3==1.25.8
Werkzeug==1.0.1
wordcloud==1.6.0
WTForms==2.2.1
